+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveService.java
@@ -426,10 +426,9 @@ public class MoveService {
       Move move,
       boolean isAutomaticReconcile,
       boolean isAutomaticAccounting,
-      boolean isUnreconcileOriginalMove)
+      boolean isUnreconcileOriginalMove,
+      LocalDate dateOfReversion)
       throws AxelorException {
-
-    LocalDate todayDate = move.getDate();
 
     Move newMove =
         moveCreateService.createMove(
@@ -437,7 +436,7 @@ public class MoveService {
             move.getCompany(),
             move.getCurrency(),
             move.getPartner(),
-            todayDate,
+            dateOfReversion,
             move.getPaymentMode(),
             MoveRepository.TECHNICAL_ORIGIN_ENTRY,
             move.getIgnoreInDebtRecoveryOk(),
@@ -461,7 +460,7 @@ public class MoveService {
               moveLine.getAccount(),
               moveLine.getCurrencyAmount(),
               !isDebit,
-              todayDate,
+              dateOfReversion,
               moveLine.getCounter(),
               moveLine.getName(),
               null);
@@ -547,7 +546,8 @@ public class MoveService {
             move,
             (boolean) assistantMap.get("isAutomaticReconcile"),
             (boolean) assistantMap.get("isAutomaticAccounting"),
-            (boolean) assistantMap.get("isUnreconcileOriginalMove"));
+            (boolean) assistantMap.get("isUnreconcileOriginalMove"),
+            (LocalDate) assistantMap.get("dateOfReversion"));
     return move;
   }
 }