+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/PaymentScheduleLineBankPaymentServiceImpl.java
@@ -168,7 +168,9 @@ public class PaymentScheduleLineBankPaymentServiceImpl extends PaymentScheduleLi
     MoveLineService moveLineService = moveService.getMoveLineService();
 
     Move advanceOrPaymentMove = paymentScheduleLine.getAdvanceOrPaymentMove();
-    Move rejectionMove = moveService.generateReverse(advanceOrPaymentMove, true, true, false);
+    Move rejectionMove =
+        moveService.generateReverse(
+            advanceOrPaymentMove, true, true, false, advanceOrPaymentMove.getDate());
     rejectionMove.setRejectOk(true);
     moveValidateService.validate(rejectionMove);